+++ src/features/meetups/components/DateSelector.tsx
@@ -3,7 +3,7 @@ import DatePicker, { registerLocale } from 'react-datepicker';
 import { useTranslation } from 'react-i18next';
 import { enGB } from 'date-fns/locale/en-GB';
 import { nl } from 'date-fns/locale/nl';
-import { getHolidaysForDate } from '../../utils/holidays';
+import { getHolidaysForDate } from '../../../utils/holidays';
 
 // Register locales once when the module is loaded
 registerLocale('en-GB', enGB);
